+++ secureli/services/logging.py
@@ -7,7 +7,7 @@ from uuid import uuid4
 
 import pydantic
 
-from secureli.abstractions.pre_commit import PreCommitAbstraction, HookConfiguration
+from secureli.services.language_support import LanguageSupportService, HookConfiguration
 from secureli.repositories.secureli_config import SecureliConfigRepository
 from secureli.utilities.git_meta import current_branch_name, git_user_email, origin_url
 from secureli.utilities.secureli_meta import secureli_version
@@ -66,10 +66,10 @@ class LoggingService:
 
     def __init__(
         self,
-        pre_commit: PreCommitAbstraction,
+        language_support: LanguageSupportService,
         secureli_config: SecureliConfigRepository,
     ):
-        self.pre_commit = pre_commit
+        self.language_support = language_support
         self.secureli_config = secureli_config
 
     def success(self, action: LogAction) -> LogEntry:
@@ -79,7 +79,7 @@ class LoggingService:
         """
         secureli_config = self.secureli_config.load()
         hook_config = (
-            self.pre_commit.get_configuration(secureli_config.overall_language)
+            self.language_support.get_configuration(secureli_config.overall_language)
             if secureli_config.overall_language
             else None
         )
@@ -109,7 +109,9 @@ class LoggingService:
         hook_config = (
             None
             if not secureli_config.overall_language
-            else self.pre_commit.get_configuration(secureli_config.overall_language)
+            else self.language_support.get_configuration(
+                secureli_config.overall_language
+            )
         )
         log_entry = LogEntry(
             status=LogStatus.failure,
